detaction-api/requirements.txt
# Flask web framework
flask>=2.0.0

# Ultralytics YOLO (for object detection)
ultralytics>=8.0.0

# PyTorch and vision libraries
torch>=1.10.0
torchvision>=0.11.0

# Computer vision and image processing
opencv-python-headless>=4.5.0
pillow>=8.3.1
scikit-image>=0.18.0

# OCR and face recognition
pytesseract>=0.3.8
face-recognition>=1.3.0
face-recognition-models>=0.3.0
dlib>=19.22.0
cmake>=3.18.0  # Required for dlib compilation

# Utilities
numpy>=1.20.0
requests>=2.26.0

# Pin setuptools to avoid pkg_resources deprecation warning
setuptools<81.0.0
